chore(*): removing the registration route
