+++ src-ebleague/app/admin/admin-csgo/admin-csgo.component.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 import { getEmptyNewCSGOGroup,
   TournamentGroup,
-  TournamentRegistration,
-  WEEKLY_SCHEDULE} from '../../../../src-common/models/tournament';
+  TournamentRegistration } from '../../../../src-common/models/tournament';
 import { ApiTournamentsService } from '../../../../src-common/services/bellumgens-api.tournaments.service';
 import { environment } from '../../../../src-common/environments/environment';
 import { IDropDroppedEventArgs } from 'igniteui-angular';
+import { WEEKLY_SCHEDULE } from '../../../../src-common/models/tournament-schedule';
 
 @Component({
   selector: 'app-admin-csgo',
