+++ src-ebleague/app/app-routing.module.ts
@@ -7,7 +7,6 @@ export const routes: Routes = [
   { path: 'format', redirectTo: '/tournament/format', pathMatch: 'full' },
   { path: 'csgo', redirectTo: '/tournament/csgo', pathMatch: 'full' },
   { path: 'sc2', redirectTo: '/tournament/sc2', pathMatch: 'full' },
-  { path: 'registration', redirectTo: '/tournament/registration', pathMatch: 'full' },
   { path: 'admin', loadChildren: () => import('./admin/admin.module').then(m => m.AdminModule) },
   { path: 'tournament', loadChildren: () => import('./tournaments/tournament.module').then(m => m.TournamentModule) },
   { path: '**', component: TournamentHomeComponent }
