+++ src-ebleague/app/tournaments/tournament-routing.module.ts
@@ -12,7 +12,7 @@ const routes: Routes = [
   { path: '', component: TournamentComponent, redirectTo: 'format' },
   { path: 'format', component: TournamentFormatComponent },
   { path: 'crew', component: ProductionCrewComponent },
-  { path: 'registration', component: TournamentRegistrationComponent },
+  // { path: 'registration', component: TournamentRegistrationComponent },
   { path: 'csgo', component: TournamentCsgoComponent, data: {
       title: 'Esports Business League - CS:GO',
       twitterTitle: 'Esports бизнес лига - CS:GO',
